fix(updateLogic)

+++ Shooter/MainMenuState.cpp
@@ -23,7 +23,7 @@ CMainMenuState::~CMainMenuState()
 
 void CMainMenuState::InitBackground()
 {
-	Background.setSize(sf::Vector2f(Window->getSize().x, Window->getSize().y));
+	Background.setSize(sf::Vector2f((float) Window->getSize().x, (float)Window->getSize().y));
 
 	if (!TextureBackground.loadFromFile("asset/sprite/menu/mainMenu/wallpaper.png"))
 		throw "ERROR:MAIN_MENU_STATE::FAILED_TO_LOAD_BACKGROUND_TEXTURE";
